+++ internal/model_request/group_leave.go
@@ -1,5 +1,5 @@
 package model_request
 
 type GroupLeave struct {
-	GroupId uint `uri:"group_id"`
+	GroupId uint64 `uri:"group_id"`
 }
